requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
icalendar==5.0.11
schedule==1.2.0
lxml==5.2.2
pytz==2023.3